recommendation_backend/requirements.txt
fastapi
uvicorn
requests
cloudscraper
beautifulsoup4
readability-lxml
sqlalchemy
psycopg2-binary
redis
requests-html
sentence-transformers
lxml[html_clean]
transformers
torch
pydantic 
python-jose 
